+++ nextjs_moralis_auth/package.json
@@ -9,13 +9,18 @@
         "lint": "next lint"
     },
     "dependencies": {
+        "@moralisweb3/next": "^2.8.0",
         "axios": "^0.27.2",
         "ethers": "^5.6.9",
-        "moralis": "2.0.1",
-        "next": "^12.2.4",
+        "moralis": "^2.8.0",
+        "next": "^13.0.5",
         "next-auth": "^4.10.3",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "wagmi": "^0.6.1"
+    },
+    "devDependencies": {
+        "@types/react": "18.0.25",
+        "typescript": "4.9.3"
     }
 }
