+++ nextjs_moralis_auth/pages/api/auth/[...nextauth].ts
@@ -1,0 +1,19 @@
+import NextAuth from 'next-auth';
+import { MoralisNextAuthProvider } from '@moralisweb3/next';
+
+export default NextAuth({
+    providers: [MoralisNextAuthProvider()],
+    // adding user info to the user session object
+    callbacks: {
+        async jwt({ token, user }) {
+            if (user) {
+                token.user = user;
+            }
+            return token;
+        },
+        async session({ session, token }) {
+            (session as { user: unknown }).user = token.user;
+            return session;
+        },
+    },
+});
