--- nextjs_moralis_auth/pages/api/auth/request-message.js
@@ -1,28 +1,0 @@
-import Moralis from 'moralis';
-
-const config = {
-    domain: process.env.APP_DOMAIN,
-    statement: 'Please sign this message to confirm your identity.',
-    uri: process.env.NEXTAUTH_URL,
-    timeout: 60,
-};
-
-export default async function handler(req, res) {
-    const { address, chain, network } = req.body;
-
-    await Moralis.start({ apiKey: process.env.MORALIS_API_KEY });
-
-    try {
-        const message = await Moralis.Auth.requestMessage({
-            address,
-            chain,
-            network,
-            ...config,
-        });
-
-        res.status(200).json(message);
-    } catch (error) {
-        res.status(400).json({ error });
-        console.error(error);
-    }
-}
