--- nextjs_moralis_auth/pages/index.js
@@ -1,5 +1,0 @@
-function HomePage() {
-    return <div>Welcome to Next.js!</div>;
-}
-
-export default HomePage;
